requirements.txt
torch
safetensors
tqdm
diffusers
pillow
numpy
requests
datasets
transformers
opencv-python
torchvision
psutil
scipy
setuptools
torchmetrics
torch-fidelity